implement simple random sampling for camera rays

+++ src/camera.rs
@@ -5,7 +5,7 @@ use ncollide::math::{Point, Scalar, Vect};
 use ncollide::ray::{Ray3};
 
 pub trait Camera {
-    fn ray_from(&self, x: u32, y: u32) -> Ray3<f64>;
+    fn ray_from(&self, x: f64, y: f64) -> Ray3<f64>;
     fn look_at_z(&mut self, at: &Pnt3<f64>, up: &Vec3<f64>);
 }
 
@@ -33,10 +33,10 @@ impl PerspectiveCamera {
 }
 
 impl Camera for PerspectiveCamera {
-    fn ray_from(&self, x: u32, y: u32) -> Ray3<f64> {
+    fn ray_from(&self, x: f64, y: f64) -> Ray3<f64> {
         let viewproj = na::to_homogeneous(&self.transform) * na::inv(&self.proj.to_mat()).expect("Projection matrix is not invertible");
-        let device_x = ((x as f64 / self.width as f64) - 0.5) * 2.0;
-        let device_y = -((y as f64 / self.height as f64) - 0.5) * 2.0;
+        let device_x = ((x / self.width as f64) - 0.5) * 2.0;
+        let device_y = -((y / self.height as f64) - 0.5) * 2.0;
         let point = Pnt4::new(device_x, device_y, -1.0, 1.0);
         let h_eye = viewproj * point;
         let eye: Pnt3<f64> = na::from_homogeneous(&h_eye);
